Create README.md
This was part of my final yea project, "Developing a Prototype Smart City Solution for Dekut and its Environs".  The basic idea is to see how universities can act as pilots in developing smart cities. The link to the result can be found at bruce.cyenite.com
bruce.cyenite.com